+++ map-positioning/src/main/java/com/getwemap/example/map/positioning/Config.kt
@@ -2,6 +2,7 @@ package com.getwemap.example.map.positioning
 
 import android.content.Context
 import androidx.preference.PreferenceManager
+import com.getwemap.sdk.core.CoreConstants
 import com.getwemap.sdk.map.helpers.MapConstants
 
 object Config {
@@ -11,10 +12,16 @@ object Config {
         with(MapConstants) {
             SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS = prefs.getBoolean(PreferenceKey.SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS.name, true)
         }
+        with(CoreConstants) {
+            USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED = prefs.getBoolean(PreferenceKey.USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED.name, true)
+            ITINERARY_RECALCULATION_ENABLED = prefs.getBoolean(PreferenceKey.ITINERARY_RECALCULATION_ENABLED.name, true)
+        }
     }
 }
 
 enum class PreferenceKey {
     // Global
-    SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS
+    SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS,
+    USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED,
+    ITINERARY_RECALCULATION_ENABLED,
 }