+++ map/src/main/java/com/getwemap/example/map/Config.kt
@@ -3,6 +3,7 @@ package com.getwemap.example.map
 import android.content.Context
 import android.graphics.Color
 import androidx.preference.PreferenceManager
+import com.getwemap.sdk.core.CoreConstants
 import com.getwemap.sdk.core.navigation.NavigationOptions
 import com.getwemap.sdk.map.helpers.MapConstants
 import com.getwemap.sdk.map.itineraries.ItineraryOptions
@@ -14,9 +15,9 @@ object Config {
     fun globalNavigationOptions(context: Context): NavigationOptions {
         val prefs = PreferenceManager.getDefaultSharedPreferences(context)
         return NavigationOptions(
-            NavigationOptions.StopOptions(stopDistanceThreshold = prefs.getString(PreferenceKey.STOP_DISTANCE_THRESHOLD.name, "15")!!.toFloat()),
-            prefs.getString(PreferenceKey.USER_POSITION_THRESHOLD.name, "25")!!.toFloat(),
-            prefs.getString(PreferenceKey.NAVIGATION_RECALCULATION_TIME_INTERVAL.name, "5")!!.toLong(),
+            arrivedDistanceThreshold = prefs.getString(PreferenceKey.ARRIVED_DISTANCE_THRESHOLD.name, "15")!!.toFloat(),
+            userPositionThreshold = prefs.getString(PreferenceKey.USER_POSITION_THRESHOLD.name, "25")!!.toFloat(),
+            navigationRecalculationTimeInterval = prefs.getString(PreferenceKey.NAVIGATION_RECALCULATION_TIME_INTERVAL.name, "5")!!.toLong(),
         )
     }
 
@@ -33,15 +34,21 @@ object Config {
         with(MapConstants) {
             SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS = prefs.getBoolean(PreferenceKey.SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS.name, true)
         }
+        with(CoreConstants) {
+            USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED = prefs.getBoolean(PreferenceKey.USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED.name, true)
+            ITINERARY_RECALCULATION_ENABLED = prefs.getBoolean(PreferenceKey.ITINERARY_RECALCULATION_ENABLED.name, true)
+        }
     }
 }
 
 enum class PreferenceKey {
     // Global
     SWITCH_LEVELS_AUTOMATICALLY_ON_USER_MOVEMENTS,
+    USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED,
+    ITINERARY_RECALCULATION_ENABLED,
 
     // Navigation
-    STOP_DISTANCE_THRESHOLD,
+    ARRIVED_DISTANCE_THRESHOLD,
     USER_POSITION_THRESHOLD,
     NAVIGATION_RECALCULATION_TIME_INTERVAL
 }